+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/features/NbtDungeon.java
@@ -55,6 +55,7 @@ public class NbtDungeon extends Feature<NbtDungeonConfig>{
     @Override
     public boolean generate(StructureWorldAccess world, ChunkGenerator chunkGenerator, Random random, BlockPos position, NbtDungeonConfig config) {
 
+        position = position.up(-1);
         Identifier nbtRL = GeneralUtils.getRandomEntry(config.nbtResourcelocationsAndWeights, random);
 
         StructureManager structureManager = world.toServerWorld().getStructureManager();
@@ -79,13 +80,13 @@ public class NbtDungeon extends Feature<NbtDungeonConfig>{
         int xMax = halfLengths.getX();
         int zMin = -halfLengths.getZ();
         int zMax = halfLengths.getZ();
-        int validOpenings = 0;
+        int wallOpenings = 0;
         int ceilingOpenings = 0;
         int ceiling = template.getSize().getY();
 
         for (int x = xMin; x <= xMax; ++x) {
             for (int z = zMin; z <= zMax; ++z) {
-                for (int y = -2; y <= ceiling + 1; ++y) {
+                for (int y = 0; y <= ceiling; ++y) {
                     mutable.set(position).move(x, y, z);
                     if(mutable.getX() >> 4 != cachedChunk.getPos().x || mutable.getZ() >> 4 != cachedChunk.getPos().z)
                         cachedChunk = world.getChunk(mutable);
@@ -98,7 +99,7 @@ public class NbtDungeon extends Feature<NbtDungeonConfig>{
                     }
                     // Floor must be complete
                     else if(!GeneralUtils.isFullCube(world, mutable, state)){
-                        if (y == -1 && !state.getMaterial().isSolid()) {
+                        if (y == 0 && !state.getMaterial().isSolid()) {
                             return false;
                         }
                         else if(state.isIn(BlockTags.LEAVES)){
@@ -110,36 +111,36 @@ public class NbtDungeon extends Feature<NbtDungeonConfig>{
                     }
 
                     // Check only along wall bottoms for openings
-                    if ((x == xMin || x == xMax || z == zMin || z == zMax) && y == 0 && isValidNonSolidBlock(config, state))
+                    if ((x == xMin || x == xMax || z == zMin || z == zMax) && y == 1 && isValidNonSolidBlock(config, state))
                     {
-                        BlockState aboveState = cachedChunk.getBlockState(mutable.move(Direction.UP));
+                        BlockState aboveState = cachedChunk.getBlockState(mutable);
                         if(config.airRequirementIsNowWater ?
                             !aboveState.getFluidState().isEmpty() :
                             aboveState.isAir())
                         {
-                            validOpenings++;
+                            wallOpenings++;
                         }
-                        mutable.move(Direction.DOWN);
                     }
 
                     // Too much open space. Quit
-                    if(validOpenings > config.maxAirSpace || ceilingOpenings >= config.maxAirSpace){
+                    if(wallOpenings > config.maxAirSpace || ceilingOpenings > config.maxAirSpace){
                         return false;
                     }
                 }
             }
         }
 
-        // offset the dungeon such as ocean dungeons down 1
-        position = position.up(config.structureYOffset);
-
         // Check if we meet minimum for open space.
-        if (validOpenings >= config.minAirSpace) {
+        if (wallOpenings >= config.minAirSpace) {
+
+            // offset the dungeon such as ocean dungeons down 1
+            position = position.up(config.structureYOffset);
+
             //RepurposedStructures.LOGGER.log(Level.INFO, nbtRL + " at X: "+position.getX() +", "+position.getY()+", "+position.getZ());
             StructurePlacementData placementsettings = (new StructurePlacementData()).setRotation(rotation).setPosition(halfLengths).setIgnoreEntities(false);
             Optional<StructureProcessorList> processor = world.toServerWorld().getServer().getRegistryManager().get(Registry.PROCESSOR_LIST_WORLDGEN).getOrEmpty(config.processor);
             processor.orElse(StructureProcessorLists.EMPTY).getList().forEach(placementsettings::addProcessor); // add all processors
-            addBlocksToWorld(template, world, chunkGenerator, mutable.set(position).move(-halfLengths.getX(), -1, -halfLengths.getZ()), placementsettings, 2, random, config);
+            addBlocksToWorld(template, world, chunkGenerator, mutable.set(position).move(-halfLengths.getX(), 0, -halfLengths.getZ()), placementsettings, 2, random, config);
             spawnLootBlocks(world, random, position, config, fullLengths, halfLengths, mutable);
             return true;
         }
@@ -226,10 +227,26 @@ public class NbtDungeon extends Feature<NbtDungeonConfig>{
                             // Set chest to face away from wall.
                             world.setBlockState(mutable, lootBlock, 2);
                             LootableContainerBlockEntity.setLootTable(world, random, mutable, config.chestResourceLocation);
+                            mutable.move(Direction.DOWN);
+                            if(lootBlock.getBlock() == Blocks.SHULKER_BOX){
+                                world.setBlockState(mutable, Blocks.SPAWNER.getDefaultState(), 2);
+                                BlockEntity blockEntity = world.getBlockEntity(mutable);
+                                if (blockEntity instanceof MobSpawnerBlockEntity) {
+                                    EntityType<?> entity = RepurposedStructures.mobSpawnerManager.getSpawnerMob(config.rsSpawnerResourcelocation, random);
 
-                            BlockState blockBelow = world.getBlockState(mutable.down());
-                            if(blockBelow.contains(SlabBlock.TYPE)){
-                                world.setBlockState(mutable.down(), blockBelow.with(SlabBlock.TYPE, SlabType.DOUBLE), 3);
+                                    if(entity != null){
+                                        ((MobSpawnerBlockEntity) blockEntity).getLogic().setEntityId(entity);
+                                    }
+                                    else{
+                                        RepurposedStructures.LOGGER.log(Level.WARN, "EntityType in a dungeon does not exist in registry!");
+                                    }
+                                }
+                            }
+                            else{
+                                BlockState blockBelow = world.getBlockState(mutable);
+                                if(blockBelow.contains(SlabBlock.TYPE)){
+                                    world.setBlockState(mutable, blockBelow.with(SlabBlock.TYPE, SlabType.DOUBLE), 3);
+                                }
                             }
 
                             currentChestCount++;